Reworked object structure.
JsUndefined and JsNull don't inherit from JsObject anymore. JsObject now all have a JsGlobal to cut down on the number of times the global is passed and the scope we require. JsGlobal now doesn't inherit from JsObject anymore, but instead there is a separate, hidden, JsObject implementation that provides the global scope. Performance improve for fannkuck is 25%.