Implemented support for string building.
Long strings are now automatically converted to JsString objects which allow for string building. Also added fast paths to operations. When an operation will probably be performed on a number, and the arguments are object, a check is done whether the argument is a double and, if so, a fast path is taken.

Bug: #42